TCD: cambios menores y arreglo de frmAuditoriaLogins

+++ MODELO/CLIENTE.cs
@@ -13,12 +13,12 @@ namespace MODELO
             ventas = new HashSet<MODELO.VENTA>();
         }
         public Int32 codigoCliente { get; set; }
-        public Int32 dni { get; set; }
+        public Int64 dni { get; set; }
         public string nombreApellido { get; set; }
         public DateTime fechaNacimiento { get; set; }
         public string direccion { get; set; }
         public LOCALIDAD localidad { get; set; }
-        public Int32 telefono { get; set; }
+        public Int64 telefono { get; set; }
         public string mail { get; set; }
         public virtual ICollection<MODELO.VENTA> ventas { get; set; }
